[change] Moved config download api views
